Correcting spelling mistakes
Aside from obvious mistakes, like the 'solution' one, several
"kata/Kata" spelling were corrected.

According to dictionaries the word is "kata" and its plural is "katas".
In general a Kata about coding is called "code kata" or "coding kata".
So all occurrences of this noun were corrected.

The unchanged occurrences are either a name (the page 'CodeKata'
of D. Thomas) or mistakes inside a quote which must not be corrected
as the quote is then falsified. For example the quote "Code Kata is an
attempt to bring..." is a mistake by D. Thomas himself because he
references to his own page, so the correct spelling would be "CodeKata"
like in the title of the page or the blog posts.

Issue: #10